Packages are installed later, stability improvements, updated README
Chocolatey and packages are now installed later because of conflicts happening when Chocolatey was trying its install .NET dependency and other big packages like Office365 while also trying to update Windows. The --background is now implemented as a background job so when specified it does not have the side effect of throwing timings off. Timings have also been generously increased and removed to lessen issues. Also found a couple of artifacts that were probably the cause for some unsucessful installations which have now been fixed.